Code pour reproduire le problème

+++ code/Robot Balance 2023/lib/StepperNB/src/StepperNB.h
@@ -8,7 +8,7 @@ public:
     // constructors:
     StepperNB(int pin_direction, int pin_step, int pin_ms1, int pin_ms2, int pin_ms3, int number_of_steps, bool direction_inverted);
 
-    void setSpeed(int target_speed_degrees_per_second);
+    void setSpeed(float target_speed_degrees_per_second);
     void setDirection(int direction);
     void setRatio(int ratio);
 
@@ -32,7 +32,7 @@ private:
     int position_steps; // microsteps = 16 x steps
     int position_degrees;
 
-    int target_speed_degrees_per_second;
+    float target_speed_degrees_per_second;
 
     int timer_period;
 };
